style: refactor code and make sure each match quit
